Little fix delta_time: If we try to pass GetTickCount() to a float variable, there is data loss and delta time will not be correctly calculated. So It is necessary to make old and new delta long variables.

+++ Tutorial10_GraphicsEngine_TransformMatrix/AppWindow.h
@@ -39,8 +39,8 @@ private:
 	PixelShader* m_ps;
 	ConstantBuffer* m_cb;
 private:
-	float m_old_delta;
-	float m_new_delta;
+	long m_old_delta;
+	long m_new_delta;
 	float m_delta_time;
 
 	float m_delta_pos;
